[#130]Fix file name and path
