+++ src/pages/index.js
@@ -6,7 +6,7 @@ import LinksToTheWork from '../components/LinksToTheWork'
 import ReleaseNote from '../components/ReleaseNotes'
 import Seo from '../components/Seo'
 import TopTitle from '../components/TopTitle'
-import FigmaGraph from '../components/FigmaGraph2'
+import FigmaGraph from '../components/FigmaGraph'
 
 import { css } from '@emotion/core'
 import GlobalStyle from '../styles/GlobalStyle'
